Fixing Azure factory
