+++ .gitignore
@@ -7,3 +7,4 @@ drivers/deployment_drivers/azure_vm/debug.xml
 debug_info.txt
 *debug.xml
 package/cloudshell_cp_azure.egg-info/*
+drivers/killPython.bat
